requirements.txt
Flask==1.0.2
SQLAlchemy==1.2.12
Flask-WTF==0.14.2
flask-socketio==3.0.1
APScheduler==2.1.2
RPi.GPIO==0.6.3
smbus2==0.1.5
SimpleCV==1.3
w1thermsensor==1.0.4
-e git+https://github.com/mapler/DHT11_Python@feature/package#egg=dht11_python
